Create new code files
I created the html and css files to write and style my code in.
Since the goal is building a single web page, I see no reason in
adding sub-directories to house the code files. Besides, it is
advisible to place the home page (index.html) in the parent directory.